Moved app_strings to helpers.make_build
