+++ corehq/apps/app_manager/suite_xml/sections/resources.py
@@ -3,7 +3,7 @@ from __future__ import unicode_literals
 
 from distutils.version import LooseVersion
 
-from corehq.apps.app_manager import id_strings
+from corehq.apps.app_manager.helpers.make_build import id_strings
 from corehq.apps.app_manager.suite_xml.contributors import SectionContributor
 from corehq.apps.app_manager.suite_xml.xml_models import LocaleResource, XFormResource, \
     PracticeUserRestoreResource, ReleaseInfoXFormResource
